backend/requirements.txt
python-dotenv
pdfplumber
Flask
Flask-Cors
numpy
openai==1.3.3
pandas
redis==5.0.1
SQLAlchemy
